Autoformatting and cleaning up code

+++ plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/analysis/cnf/generator/configuration/util/MIGComparator.java
@@ -93,7 +93,7 @@ public class MIGComparator implements Comparator<LiteralSet> {
 				vOut += (info.strongOut) + info.weakOut;
 			}
 		}
-		return (double) (vIn - (vOut * vOut));
+		return vIn - (vOut * vOut);
 	}
 
 	public int getOut(LiteralSet... set) {
